+++ tests/Crypt/CryptTest.php
@@ -2,26 +2,27 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Security\Tests;
+namespace Yiisoft\Security\Tests\Crypt;
 
-require_once __DIR__ . '/MockHelper.php';
-
+use Exception;
 use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\RequiresPhpExtension;
+use RuntimeException;
 use Yiisoft\Security\AuthenticationException;
 use Yiisoft\Security\Crypt;
 use Yiisoft\Security\MockHelper;
+use Yiisoft\Security\Tests\TestCase;
 
+use function dirname;
+
+require_once dirname(__DIR__) . '/MockHelper.php';
+
+#[RequiresPhpExtension('openssl')]
 final class CryptTest extends TestCase
 {
     protected function tearDown(): void
     {
         MockHelper::resetMocks();
-    }
-
-    private function getCrypt(): Crypt
-    {
-        // speed up test running
-        return (new Crypt())->withDerivationIterations(1000);
     }
 
     public function testEncryptAndDecryptByPassword(): void
@@ -32,7 +33,6 @@ final class CryptTest extends TestCase
 
         $encryptedData = $crypt->encryptByPassword($data, $key);
         $decryptedData = $crypt->decryptByPassword($encryptedData, $key);
-
 
         $this->assertNotSame($data, $encryptedData);
         $this->assertEquals($data, $decryptedData);
@@ -409,7 +409,7 @@ final class CryptTest extends TestCase
      * @param string $data plaintext hex string
      * @param string $encrypted ciphertext hex string
      *
-     * @throws \Exception
+     * @throws Exception
      */
     #[DataProvider('dataProviderEncryptByKeyCompatibilityWithMcrypt')]
     public function testEncryptByKeyCompatibilityWithMcrypt(string $key, string $data, string $encrypted): void
@@ -428,7 +428,7 @@ final class CryptTest extends TestCase
      * @param string $data plaintext hex string
      * @param string $encrypted ciphertext hex string
      *
-     * @throws \Exception
+     * @throws Exception
      */
     #[DataProvider('dataProviderEncryptByKeyCompatibilityWithOpenSsl')]
     public function testEncryptByKeyCompatibilityWithOpenSsl(string $key, string $data, string $encrypted): void
@@ -747,13 +747,13 @@ final class CryptTest extends TestCase
      * @param string $data plaintext hex string
      * @param string $encrypted ciphertext hex string
      *
-     * @throws \Exception
+     * @throws Exception
      */
     #[DataProvider('dataProviderEncryptByPasswordMcryptCompatibility')]
     public function testEncryptByPasswordCompatibilityWithMcrypt(
         string $password,
         string $data,
-        string $encrypted
+        string $encrypted,
     ): void {
         $crypt = $this->getCrypt();
 
@@ -768,13 +768,13 @@ final class CryptTest extends TestCase
      * @param string $data plaintext hex string
      * @param string $encrypted ciphertext hex string
      *
-     * @throws \Exception
+     * @throws Exception
      */
     #[DataProvider('dataProviderEncryptByPasswordOpenSslCompatibility')]
     public function testEncryptByPasswordCompatibilityWithOpenSsl(
         string $password,
         string $data,
-        string $encrypted
+        string $encrypted,
     ): void {
         $crypt = $this->getCrypt();
 
@@ -784,24 +784,16 @@ final class CryptTest extends TestCase
         $this->assertEquals($data, $crypt->decryptByPassword($encrypted, $password));
     }
 
-    public function testOpensslNotLoadedException(): void
-    {
-        $this->expectException(\RuntimeException::class);
-
-        MockHelper::$mock_extension_loaded = false;
-        $this->getCrypt();
-    }
-
     public function testUnknownCipher(): void
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(RuntimeException::class);
 
         new Crypt('test');
     }
 
     public function testNotAllowedCypher(): void
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(RuntimeException::class);
 
         new Crypt('BF-CBC');
     }
@@ -832,7 +824,7 @@ final class CryptTest extends TestCase
 
     public function testOpensslEncryptException(): void
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(RuntimeException::class);
 
         MockHelper::$mock_openssl_encrypt = false;
         $crypt = $this->getCrypt();
@@ -841,11 +833,17 @@ final class CryptTest extends TestCase
 
     public function testOpensslDecryptException(): void
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(RuntimeException::class);
 
         MockHelper::$mock_openssl_decrypt = false;
         $crypt = $this->getCrypt();
         $encrypted = $crypt->encryptByPassword('test', 'test');
         $crypt->decryptByPassword($encrypted, 'test');
+    }
+
+    private function getCrypt(): Crypt
+    {
+        // speed up test running
+        return (new Crypt())->withDerivationIterations(1000);
     }
 }